+++ packages/graphql-hooks/src/index.js
@@ -1,5 +1,7 @@
 import ClientContext from './ClientContext'
 import GraphQLClient from './GraphQLClient'
+import LocalGraphQLClient from './LocalGraphQLClient'
+import LocalGraphQLError from './LocalGraphQLError'
 import useClientRequest from './useClientRequest'
 import useQuery from './useQuery'
 import useSubscription from './useSubscription'
@@ -13,6 +15,8 @@ const useMutation = (query, options) =>
 export {
   ClientContext,
   GraphQLClient,
+  LocalGraphQLClient,
+  LocalGraphQLError,
   useClientRequest,
   useQuery,
   useSubscription,