+++ examples/full-ws-transport/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.4.0](https://github.com/nearform/graphql-hooks/compare/graphql-hooks-full-ws-transport-example@0.3.0...graphql-hooks-full-ws-transport-example@0.4.0) (2022-03-25)
+
+**Note:** Version bump only for package graphql-hooks-full-ws-transport-example
+
+
+
+
+
+# [0.3.0](https://github.com/nearform/graphql-hooks/compare/graphql-hooks-full-ws-transport-example@0.2.1...graphql-hooks-full-ws-transport-example@0.3.0) (2022-03-25)
+
+**Note:** Version bump only for package graphql-hooks-full-ws-transport-example
+
+
+
+
+
 ## [0.2.1](https://github.com/nearform/graphql-hooks/compare/graphql-hooks-full-ws-transport-example@0.2.0...graphql-hooks-full-ws-transport-example@0.2.1) (2022-03-11)
 
 **Note:** Version bump only for package graphql-hooks-full-ws-transport-example